requirements.txt
biopython
numpy
pandas
scikit-learn
torch
transformers
matplotlib
peft
pytorch_metric_learning